Refactor imports to use domain datamodels for CommandType and MessageStatus across multiple files
