+++ integrations/discord/parser.py
@@ -9,7 +9,8 @@ from discord.channel import TextChannel
 
 import libs.global_value as g
 from integrations.base.interface import MessageParserDataMixin, MessageParserInterface
-from integrations.protocols import CommandType, MsgData, PostData, StatusData
+from integrations.protocols import MsgData, PostData, StatusData
+from libs.domain.datamodels import CommandType
 
 if TYPE_CHECKING:
     from integrations.discord.adapter import ServiceAdapter
