+++ libs/commands/ranking/rating.py
@@ -7,9 +7,8 @@ from typing import TYPE_CHECKING
 import pandas as pd
 
 import libs.global_value as g
-from integrations.protocols import CommandType
 from libs.data import aggregate, loader
-from libs.domain.datamodels import GameInfo
+from libs.domain.datamodels import CommandType, GameInfo
 from libs.functions import message
 from libs.functions.compose import badge
 from libs.types import StyleOptions
