+++ libs/commands/registry/team.py
@@ -6,9 +6,9 @@ import logging
 from typing import TYPE_CHECKING, TypedDict, cast
 
 import libs.global_value as g
-from integrations.protocols import CommandType
 from libs.bootstrap.app_config import BaseSection
 from libs.data import initialization, loader, modify
+from libs.domain.datamodels import CommandType
 from libs.utils import dbutil, formatter, textutil, validator
 
 if TYPE_CHECKING:
